move apply method to seperate package
